+++ src/app/modules/rentals/pages/view-reservation/view-reservation.component.html
@@ -192,14 +192,14 @@
 <app-modal [(show)]="show_assign_serials">
 	<h2>Asignar Inventario</h2>
 	<div style="min-height:400px">
-		<div class="row">
-			<!--app-code-reader (onDetect)="addSerial($event)"></app-code-reader-->
-			<div class="col-12 col-md-6">
+		<div class="row" *ngIf="show_assign_serials">
+			<app-code-reader (onDetect)="onCodeArrived($event)"></app-code-reader>
+			<!--div class="col-12 col-md-6">
 				<input type="text" class="form-control" placeholder="Buscar..." name="search_serials"  [(ngModel)]="search_serials">
 			</div>
 			<div class="col-12 col-md-6">
 				<button class="btn btn-primary" (click)="addSerial(search_serials)">Añadir</button>
-			</div>
+			</div-->
 		</div>
 	</div>
 </app-modal>
