+++ src/app/modules/shared/BuildInfo.ts
@@ -1,1 +1,1 @@
-export const BuildInfo = {timestamp:1739410170000}
+export const BuildInfo = {timestamp:1739927108000}
